% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_const_matrix}
\alias{get_const_matrix}
\title{Construct constraint matrix for LP formulation}
\usage{
get_const_matrix(n)
}
\arguments{
\item{n}{number of counterparties}
}
\value{
Matrix of size \eqn{n \times n(n-1)}
}
\description{
Construct constraint matrix for LP formulation
}
