% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_2_vec}
\alias{mat_2_vec}
\title{Matrix to vector representation}
\usage{
mat_2_vec(X)
}
\arguments{
\item{X}{a square matrix of size \eqn{n \times n}}
}
\value{
A vector of length \eqn{n(n-1)}
}
\description{
Matrix to vector representation
}
