% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_min_TX}
\alias{solve_min_TX}
\title{Minimize number of payments using linear programming}
\usage{
solve_min_TX(D = NULL, b = NULL)
}
\arguments{
\item{D}{matrix of payments between members of size \eqn{n \times n} with
nonnegative entries}

\item{b}{vector of size \eqn{n} with net income of each member (positive or
negative). Entries of \code{b} must add up to 0. If both \code{b} and \code{D} are
present, \code{b} is used}
}
\value{
Matrix of payments of size \eqn{n \times n} with the least number
of nonzero entries
}
\description{
Minimize number of payments using linear programming
}
\examples{
solve_min_TX(D = matrix(rexp(5*5, 1/10), nrow = 5))

}
