% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_2_mat}
\alias{vec_2_mat}
\title{Vector to matrix representation}
\usage{
vec_2_mat(x)
}
\arguments{
\item{x}{a vector of length \eqn{n(n-1)}}
}
\value{
A square matrix of size \eqn{n \times n}
}
\description{
Vector to matrix representation
}
